chore: Lint

+++ iOS/Catalog/Catalog/ExternalColorSelection/Data/ExteriorColorResponseDTO.swift
@@ -10,7 +10,7 @@ import Foundation
 struct ExteriorColorResponseDTO: Decodable {
   let id: Int?
   let name: String?
-  let choiceRatio: Int?
+  let choiceRatio: Double?
   let price: Int?
   let hexCode: String?
   let images: [String]?
